backend/agents/requirements_cgp_agent.txt
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.3.0
langchain-chroma>=1.0.0
langchain-text-splitters>=0.3.0
chromadb>=1.0.0
pypdf>=4.0.0
ddgs>=8.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
anthropic>=0.39.0
